requirements.txt
datasets
pandas
numpy
scikit-learn
rouge-score
sentence-transformers
nano-graphrag
dashscope
google-genai
openai
tiktoken
requests
